change dollarvalue and split overview:
