+++ src/components/custom/index.ts
@@ -1,9 +1,20 @@
+//Amount Change
+export * from '@/components/custom/AmountChange/AmountChange';
+
+// AmountDisplay
+export * from '@/components/custom/AmountDisplay/AmountDisplay';
+
 // Card
 export * from '@/components/custom/Card/Card';
+
 export * from '@/components/custom/Card/CardContent';
+
 export * from '@/components/custom/Card/CardFooter';
+
 export * from '@/components/custom/Card/CardHeader';
+
 export * from '@/components/custom/Card/CardHeaderTabsList';
+
 export * from '@/components/custom/Card/CardHeaderTabsTrigger';
 
 // Color
@@ -12,10 +23,6 @@ export * from '@/components/custom/Color/Color';
 // Distribution Bar
 export * from '@/components/custom/DistributionBar/DistributionBar';
 export * from '@/components/custom/DistributionBar/DistributionBarSegment';
-
-// DollarValue
-export * from '@/components/custom/DollarValue/DollarValue';
-export * from '@/components/custom/DollarValue/DollarValueChange';
 
 // Glossary
 export * from '@/components/custom/Glossary/Glossary';
